First version that works
The library and console application build. The console application is
simple and provides no user interaction. Its purpose is to help a user
get started with the library.
