+++ .gitignore
@@ -3,6 +3,9 @@
 ##
 ## Get latest from https://github.com/github/gitignore/blob/main/VisualStudio.gitignore
 
+# Project specific files
+/VodConsole/examples/**
+
 # User-specific files
 *.rsuser
 *.suo
