+++ ValidateOscalDocuments.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VodLibrary", "ValidateOscalDocuments\VodLibrary.csproj", "{4801A327-8D32-4880-AB3A-4EF15DA8DEF7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VodLibrary", "VodLibrary\VodLibrary.csproj", "{4801A327-8D32-4880-AB3A-4EF15DA8DEF7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VodConsole", "VodConsole\VodConsole.csproj", "{69002B99-CB17-4703-88B5-47A3F0138AEE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VodConsole", "VodConsole\VodConsole.csproj", "{69002B99-CB17-4703-88B5-47A3F0138AEE}"
 	ProjectSection(ProjectDependencies) = postProject
 		{4801A327-8D32-4880-AB3A-4EF15DA8DEF7} = {4801A327-8D32-4880-AB3A-4EF15DA8DEF7}
 	EndProjectSection
