+++ VodLibrary/schema/v1.0.4/oscal_catalog_schema.xsd
